fix norm

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: hkhalil <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/08 16:02:17 by hkhalil           #+#    #+#             */
-/*   Updated: 2021/11/10 19:27:00 by hkhalil          ###   ########.fr       */
+/*   Updated: 2021/11/12 17:42:32 by hkhalil          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,9 +22,9 @@ int	ft_isspace(int c)
 
 int	ft_atoi(const char *nptr)
 {
-	int			sign;
-	int			i;
-	long long	n;
+	int	sign;
+	int	i;
+	int	n;
 
 	sign = 1;
 	i = 0;
